requirements.txt
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
langchain-core>=0.3.0
slack-bolt>=1.23.0
python-dotenv>=1.0.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pathlib
pickleshare
typing
uuid